+++ js/data.js
@@ -1,3 +1,19 @@
-const str = 'I am a Keks cat.';
+import { getRandInt } from './util.js';
 
-export {str};
+function generateNewPosts () {
+  const photos = [];
+  for(let i = 1; i < 26; i++) {
+    photos.push(
+      {
+        id: i,
+        url: `photos/${i}.jpg`,
+        description: `my best photo ${i}`,
+        likes: getRandInt(15, 200),
+        comments: getRandInt(0, 200)
+      }
+    );
+  }
+  return photos;
+}
+
+export {generateNewPosts};
