+++ js/main.js
@@ -1,6 +1,5 @@
-import { lengthCheck, generateNewPosts } from './util.js';
-import {str} from './data.js';
+import { getRandInt, lengthCheck } from './util.js';
+import './data.js';
 
-lengthCheck(str, 4);
-
-generateNewPosts();
+getRandInt(0, 200);
+lengthCheck('', 12);
